extracted GraphQLCanadianProvince.kt from GraphQLProvince.kt
